% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-utils.R, R/utils.R
\name{apply_dim_reduct}
\alias{apply_dim_reduct}
\title{FUNCTION_TITLE}
\usage{
apply_dim_reduct(
  data_list,
  dim = NULL,
  mode = "FSM",
  random_seed = 42,
  upsample = FALSE
)

apply_dim_reduct(
  data_list,
  dim = NULL,
  mode = "FSM",
  random_seed = 42,
  upsample = FALSE
)
}
\arguments{
\item{data_list}{\link{(a,n)...(z,n)}  list of datasets}

\item{dim}{desired dimension after dimensionality reduction}

\item{mode}{method for finding anchors, default on 'FSM'}

\item{random_seed}{PARAM_DESCRIPTION, Default: 42}

\item{upsample}{flag for upsample}
}
\value{
OUTPUT_DESCRIPTION

data_list   \link{(a,dim)...(z,dim)} list of datasets
}
\description{
FUNCTION_DESCRIPTION

The function reduces the datasets to dim subspaces
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[reticulate]{import}}, \code{\link[reticulate]{np_array}}, \code{\link[reticulate]{py_run}}
\code{\link[glue]{glue}}
}
