% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_marker_gene}
\alias{calculate_marker_gene}
\title{calculate_marker_gene}
\usage{
calculate_marker_gene(data_matrix, labels_pred, topn, gene_label)
}
\arguments{
\item{labels_pred}{predicted labels}

\item{topn}{}

\item{gene_label}{Out:}

\item{data}{raw data (cell*feature)}
}
\value{
gene_df_figure  \link{gene_df,figure}
}
\description{
calculate_marker_gene
}
