% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_lineage}
\alias{compute_lineage}
\title{compute_lineage}
\usage{
compute_lineage(ZW, labels_combined, root_cluster, name, reverse)
}
\arguments{
\item{ZW}{OCAT features}

\item{labels_combined}{true labels}

\item{root_cluster}{}

\item{name}{}

\item{reverse}{Out:}
}
\value{
Lineage  \link{Lineage, root_cluster, cluster_labels, tree}
}
\description{
The function infers Lineages over clusters with the OCAT features, predicted/true cluster labels and a user-specified root_cluster.
}
