% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_ptime}
\alias{compute_ptime}
\title{compute_ptime}
\usage{
compute_ptime(ZW, labels_combined, Lineage, root_cluster, root_cell = NULL)
}
\arguments{
\item{ZW}{OCAT features}

\item{labels_combined}{true labels}

\item{Lineage}{OCAT Lineage}

\item{root_cluster}{}

\item{root_cell}{Out:}
}
\value{
P  \link{Ptime, root_cell_list}
}
\description{
The function infers pseudotime for individual cell using the OCAT extracted features and the predicted lineage.
}
