% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-utills.R
\name{dim_estimate}
\alias{dim_estimate}
\title{FUNCTION_TITLE}
\usage{
dim_estimate(data_list)
}
\arguments{
\item{data_list}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
