% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-utills.R, R/utills.R
\name{evaluate_clusters}
\alias{evaluate_clusters}
\title{FUNCTION_TITLE}
\usage{
evaluate_clusters(ZW, num_clusters)

evaluate_clusters(ZW, num_clusters)
}
\arguments{
\item{ZW}{OCAT feature matrix}

\item{num_clusters}{num of clusters
Out:}
}
\value{
OUTPUT_DESCRIPTION

labels_pred  predicted labels
}
\description{
FUNCTION_DESCRIPTION

The function wraps the evaluate_cluster function from OCAT python packagee
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[reticulate]{import}}, \code{\link[reticulate]{np_array}}
}
