% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-utils.R, R/utils.R
\name{find_anchors}
\alias{find_anchors}
\title{FUNCTION_TITLE}
\usage{
find_anchors(data_list, m_list)

find_anchors(data_list, m_list)
}
\arguments{
\item{data_list}{\link{(a,n)...(z,n)} list of datasets}

\item{m_list}{lift of num of anchors
Out:}
}
\value{
OUTPUT_DESCRIPTION

anchor_list \link{(m,n)...(m,n)} list of anchors
}
\description{
FUNCTION_DESCRIPTION

The function generates the corresponding anchors for the datasets
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[reticulate]{import}}, \code{\link[reticulate]{np_array}}, \code{\link[reticulate]{py_run}}
\code{\link[glue]{glue}}
}
