% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-utills.R, R/utills.R
\name{install_packages}
\alias{install_packages}
\title{FUNCTION_TITLE}
\usage{
install_packages()

install_packages()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

install required python packages
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[reticulate]{py_install}}
}
