% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-utills.R, R/utills.R
\name{l2_normalization}
\alias{l2_normalization}
\title{FUNCTION_TITLE}
\usage{
l2_normalization(data_list)

l2_normalization(data_list)
}
\arguments{
\item{data_list}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

normalize all cells by L2 normalization
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[Matrix]{colSums}}, \code{\link[Matrix]{Matrix}}
}
