% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-utils.R, R/utils.R
\name{normalize_data}
\alias{normalize_data}
\title{FUNCTION_TITLE}
\usage{
normalize_data(data_list)

normalize_data(data_list)
}
\arguments{
\item{data_list}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

normalize all cells by log10(x+!)
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[Matrix]{Matrix}}
}
