% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-utils.R, R/utils.R
\name{normalized_mutual_info_score}
\alias{normalized_mutual_info_score}
\title{FUNCTION_TITLE}
\usage{
normalized_mutual_info_score(labels_true, labels_pred)

normalized_mutual_info_score(labels_true, labels_pred)
}
\arguments{
\item{labels_true}{ground truth labels}

\item{labels_pred}{predicted labels
Out:}
}
\value{
OUTPUT_DESCRIPTION

NMI     NMI score
}
\description{
FUNCTION_DESCRIPTION

The function wraps the normalized_mutual_info_score from sklearn.metrics.cluster
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[reticulate]{import}}
}
