% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_utils.R
\name{plot_UMAP}
\alias{plot_UMAP}
\title{plot_UMAP}
\usage{
plot_UMAP(
  umap_embedding,
  labels_true,
  labels_predicted,
  legend_labels = NULL,
  title = "",
  x = "bottomright",
  y.intersp = 1
)
}
\arguments{
\item{labels_true}{true labels}

\item{labels_predicted}{predicted labels}

\item{legend_labels}{corresponding class to clustering labels}

\item{title}{UMAP embedding}

\item{x}{the x and y co-ordinates to be used to position the legend. They can be specified by keyword or in any way which is accepted by xy.coords}

\item{y.intersp}{character interspacing factor for vertical (y) line distance}

\item{embeddding}{UMAP embedding}
}
\description{
The function plot the UMAP embedding
}
