% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-utills.R, R/utils.R
\name{preprocess}
\alias{preprocess}
\title{FUNCTION_TITLE}
\usage{
preprocess(data_list, log_norm = log_norm, l2_norm = l2_norm)

preprocess(data_list, log_norm = log_norm, l2_norm = l2_norm)
}
\arguments{
\item{data_list}{\link{(a,n)...(z,n)}    -- list of datasets.}

\item{log_norm}{flag for log10 normalization.}

\item{l2_norm}{flag  for L2 normalization.
Out:}
}
\value{
OUTPUT_DESCRIPTION

data_list   \link{(a,dim)...(z,dim)}    -- list of normalized datasets
}
\description{
FUNCTION_DESCRIPTION

normalize cells expression
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
