% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rOCAT-package.R
\docType{package}
\name{rOCAT-package}
\alias{rOCAT}
\alias{rOCAT-package}
\title{rOCAT: An R Interface to OCAT}
\description{
This is the R Interface for One Cell At a Time(OCAT) python implmentation.
}
\author{
\strong{Maintainer}: Chong Shao \email{chong.shao@mail.utoronto.ca}

}
\keyword{internal}
