% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_cell_inference}
\alias{run_cell_inference}
\title{run_cell_inference}
\usage{
run_cell_inference(data_list_inf, ZW_db, labels_db, db_list)
}
\arguments{
\item{data_list_inf}{data list contains the inference data_list}

\item{ZW_db}{ZW for the reference data_list}

\item{db_list}{\link{anchor_list, s_list, W_anchor, Wm} for reference data list
Out:}

\item{labels_true_reference}{true labels for the reference list}
}
\value{
ZW_inf_labels  \link{ZW for inference data, predicted labels for inference datasets}
}
\description{
The function predicts the labels of inference data from our ZW in reference data list
}
