% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-utils.R
\name{run_OCAT.default}
\alias{run_OCAT.default}
\title{FUNCTION_TITLE}
\usage{
\method{run_OCAT}{default}(
  data_list,
  m_list = NULL,
  s_list = NULL,
  dim = NULL,
  p = 0.3,
  log_norm = TRUE,
  l2_norm = TRUE,
  if_inference = FALSE,
  random_seed = 42
)
}
\arguments{
\item{data_list}{PARAM_DESCRIPTION}

\item{m_list}{PARAM_DESCRIPTION, Default: NULL}

\item{s_list}{PARAM_DESCRIPTION, Default: NULL}

\item{dim}{PARAM_DESCRIPTION, Default: NULL}

\item{p}{PARAM_DESCRIPTION, Default: 0.3}

\item{log_norm}{PARAM_DESCRIPTION, Default: TRUE}

\item{l2_norm}{PARAM_DESCRIPTION, Default: TRUE}

\item{if_inference}{PARAM_DESCRIPTION, Default: FALSE}

\item{random_seed}{PARAM_DESCRIPTION, Default: 42}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
