% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-utils.R, R/utils.R
\name{run_OCAT}
\alias{run_OCAT}
\title{FUNCTION_TITLE}
\usage{
run_OCAT(object, group.by.vars, ...)

run_OCAT(object, group.by.vars, ...)
}
\arguments{
\item{object}{Seurat Object}

\item{group.by.vars}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}

\item{m_list}{num of anchors}

\item{s_list}{num of anchors to be selected}

\item{dim}{default values of dim}

\item{p}{percentage of NNs to consider,default = 0.3}

\item{log_norm}{num of anchors to be selected}

\item{if_inference}{flag for cell inference}

\item{random_seed}{default = 42
Out:}
}
\value{
OUTPUT_DESCRIPTION

ZW (a+...+z, m)  OCAT feature matrix
}
\description{
FUNCTION_DESCRIPTION

The function performs OCAT pipeline and return the sparsed encoding oof edges
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
