% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_utils.R
\name{run_UMAP}
\alias{run_UMAP}
\title{run_UMAP}
\usage{
run_UMAP(embeddding)
}
\arguments{
\item{embeddding}{OCAT embedding
Out:}
}
\value{
OCAT_umap calculated UMAP embedding
}
\description{
The function calculate the UMAP embedding
}
