% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-utils.R
\name{test_function}
\alias{test_function}
\title{FUNCTION_TITLE}
\usage{
test_function(data_list)
}
\arguments{
\item{data_list}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
