% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-utils.R, R/utils.R
\name{load_all}
\alias{load_all}
\title{FUNCTION_TITLE}
\usage{
load_all()

load_all()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

install required python packages
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[reticulate]{py_install}}
}
